Use Yarn Workspaces
As outlined on their blog [1]

This results in a HUGE number of failures when running `yarn start` from
`packages/static`, but in no failures whatsoever when running `yarn
start` from `packages/webapp`.

Once again, as in the previous commit, this points to `react-static`
doing something unconventional when looking up dependencies.

  [1]: https://yarnpkg.com/blog/2017/08/02/introducing-workspaces/